Changes variables p, ptcl and ptcls to ptclnum, ptclnum and ptclnums, because these were effectively "counters" or indeces of particle numbers, NOT particles themselves. These names make things clearer by being more descriptive.
Most significantly, storage of particle orientations in tomo_xforms.json happened AFTER all particles had been aligned (which was absurd).
Now they are stored AS EACH alignment task is completed. This will shortly enable "recovering" completed alignments from crashed jobs. This can potentially be a very significant time-saver for large sets, or for sets involving big particles.

Lastly, I changed the labeles for transforms in tomo_xforms.json files so that the label identifying each particle transform doesn't have an arbitrarily large/short numbers of zeros.
For example, in a set of 5 particles, a "tomo_0005" label makes no sense (why the zeros?). On the other hand, the 4 digits in the string would be insufficient for a dataset of 10,000 particles or more.
Thus, the label is NOW determined based on the size of the data set. I.e., particle #5 in a data set of 9 or less, will have a label "tomo_5", whereas particle 5 in a dataset between 1000 and 9999 will be labeled "tomo_0005".