added some test codes to automatically detect the lines where division is occuring and making them python 2 to 3 compatible.
testdivisionchange is a test file which has the same source code as sparx_fundamentals so that If i make any changes to it the orignial code is not affected.

finddivisionoperator is search code for finding all the / operators in the code and then creating independent scenarios for different cases (one time, multiple time , long equation where / was used.)

it creates a new file testdivisionchangenew which is now python 2 and python 3 compatible.  Once the code is completed then a new folder will be created with updated version of libpy source codes.