get list of e2 programs from source instead of install location to avoid running deleted programs during tests
When a program is removed from source, cmake does not remove it from installation location. So, over time, the install location will contain old files which are not expected to work anymore. For example, if a file named e2qt4.py worked with PyQt4, but was renamed to e2qt5.py after upgrading to PyQt5, e2qt4.py will remain in the installation location and running tests on all e2 files will cause e2qt4.py to fail, because it has PyQt4 imports. Instead, if programs list is populated from the /programs directory in the source, e2qt4.py will not be run, even though it is still present in the installation directory.

This is important for local development environments and build machines which are not usually set up from scratch every time the tests are run. It shouldn't affect the online CI services, because there the tests are run in clean environments.